population genetics evolution model
TurtlesModel contents all the classes that contruct the turtle object, data structure, and behavior. The TurtlesView contanin the classes and methods that build the interface, space of development of the turtles, interface to view info for each turtle, figure displaying evolution process plot and also contain the engine to run the system. The TurtlesView is built using customtkinter and tkinter for the interface.